Added exception so staff can ping techno.

+++ src/main/java/com/technovision/technobot/listeners/AutomodListener.java
@@ -2,6 +2,7 @@ package com.technovision.technobot.listeners;
 
 import com.technovision.technobot.TechnoBot;
 import com.technovision.technobot.logging.AutoModLogger;
+import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.TextChannel;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
@@ -22,22 +23,30 @@ public class AutomodListener extends ListenerAdapter {
     @Override
     public void onMessageReceived(@Nonnull MessageReceivedEvent event) {
         String message = event.getMessage().getContentRaw();
-        if (event.getAuthor().isBot()) { return; }
+        if (event.getAuthor().isBot()) {
+            return;
+        }
         if (event.getMessage().getMentionedMembers().size() > 0) {
-            if (event.getMessage().getMentionedMembers().contains(event.getGuild().getMemberById(595024631438508070L))) {
+            if (event.getMessage().getMentionedMembers().contains(event.getGuild().getMemberById(595024631438508070L)) &&
+                !hasRole(event.getMember(), "staff")) {
                 TextChannel channel = event.getGuild().getTextChannelsByName("RULES-AND-INFO", true).get(0);
                 event.getMessage().delete().queue();
                 event.getChannel().sendMessage("<@!" + event.getAuthor().getIdLong() + ">, do not ping TechnoVision! <#" + channel.getId() + ">").queue();
                 LOGGER.log(event.getGuild(), event.getTextChannel(), event.getAuthor(), event.getJDA().getSelfUser(), AutoModLogger.Infraction.PING);
             }
-        }
-        else if (message.toLowerCase().contains("discord.gg/")) {
+        } else if (message.toLowerCase().contains("discord.gg/")) {
             if (!event.getChannel().getName().equals(ADVERTISE_CHANNEL)) {
                 event.getMessage().delete().queue();
                 TextChannel channel = event.getGuild().getTextChannelsByName("RULES-AND-INFO", true).get(0);
-                event.getChannel().sendMessage("<@!" + event.getAuthor().getId() + ">, " + "you are not allowed to advertise in this channel! <#"+channel.getId()+">").queue();
+                event.getChannel().sendMessage("<@!" + event.getAuthor().getId() + ">, " + "you are not allowed to advertise in this channel! <#" + channel.getId() + ">").queue();
                 LOGGER.log(event.getGuild(), event.getTextChannel(), event.getAuthor(), event.getJDA().getSelfUser(), AutoModLogger.Infraction.INVITE);
             }
         }
+    }
+
+    public boolean hasRole(Member member, String roleName) {
+        return member.getRoles().stream()
+                       .filter(role -> role.getName().equals(roleName))
+                       .count() == 1;
     }
 }
